float on price

+++ taleoftiles/models.py
@@ -182,9 +182,9 @@ class Product(models.Model):
 class Price(models.Model):
     size    = models.ForeignKey(Tag,     blank = True, null = True, on_delete=models.SET_NULL, related_name='prices')
     product = models.ForeignKey(Product, blank = True, null = True, on_delete=models.SET_NULL, related_name='prices' )
-    euros       = models.PositiveIntegerField(default = 10)
-    m2_box      = models.PositiveIntegerField(default = 10)
-    weight_box  = models.PositiveIntegerField(default = 10)
+    euros       = models.FloatField(default = 10)
+    m2_box      = models.FloatField(default = 10)
+    weight_box  = models.FloatField(default = 10)
 
     def __str__(self):
         return self.size.name + self.product.publication.title + str(self.euros)
